scraper/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Web scraping with anti-detection
botasaurus>=4.0.0

# HTML parsing (used in google_search/helpers.py)
beautifulsoup4>=4.12.0

# Environment variables
python-dotenv>=1.0.0
